fix compile errors / warnings on Windows

+++ include/pion/tcp/connection.hpp
@@ -272,7 +272,7 @@ public:
     inline std::error_code connect(const asio::ip::address& remote_addr,
                                              const unsigned int remote_port)
     {
-        asio::ip::tcp::endpoint tcp_endpoint(remote_addr, remote_port);
+        asio::ip::tcp::endpoint tcp_endpoint(remote_addr, (unsigned short)remote_port);
         return connect(tcp_endpoint);
     }
